requirements.txt
Pillow==9.4.0
PyQt5-sip==12.13.0
PyQt5==5.15.10
PyQt5-stubs==5.15.6.0
PyQt5-Qt5==5.15.2
uic==0.0.1
